refactor: organize auxiliary service initializers
