+++ src/bootstrapper/stages/auxiliary/index.js
@@ -1,0 +1,11 @@
+// src/bootstrapper/stages/auxiliary/index.js
+
+export { initEngineUIManager } from './initEngineUIManager.js';
+export { initSaveGameUI } from './initSaveGameUI.js';
+export { initLoadGameUI } from './initLoadGameUI.js';
+export { initLlmSelectionModal } from './initLlmSelectionModal.js';
+export { initCurrentTurnActorRenderer } from './initCurrentTurnActorRenderer.js';
+export { initSpeechBubbleRenderer } from './initSpeechBubbleRenderer.js';
+export { initProcessingIndicatorController } from './initProcessingIndicatorController.js';
+
+export { initializeAuxiliaryServicesStage } from '../initializeAuxiliaryServicesStage.js';
